backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
motor>=3.3.0
pydantic[email]>=2.4.0
langchain-core>=0.1.52
langchain-google-genai>=1.0.3
pypdf>=4.0.0
certifi>=2024.0.0
python-multipart>=0.0.9
passlib>=1.7.4
bcrypt>=4.0.1,<4.1.0
PyJWT>=2.8.0
bcrypt==4.0.1
passlib==1.7.4
